refactor(datasync): move datasync resolvers to datasync plugin
