+++ packages/graphback-core/src/crud/mappingHelpers.ts
@@ -170,11 +170,3 @@ export function getInputFieldType(field: GraphQLField<any, any>): GraphQLInputTy
 
   return undefined;
 }
-
-/**
- * Get the name of delta query for a model
- * @param typeName Name of the model
- */
-export function getDeltaQuery(typeName: string) {
-  return `sync${pluralize(typeName)}`;
-}
